@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+4
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=3.4.4 <4.0.0'
 
 dependencies:
   flutter:
@@ -26,19 +26,20 @@ dependencies:
   flutter_riverpod: ^2.3.5
   flutter_dotenv: ^5.0.2
   dotenv: ^4.1.0
-  photo_view: ^0.14.0
+  photo_view: ^0.15.0
   http: ^0.13.4
-  intl: ^0.18.1
-  file: ^6.1.4
+  intl: ^0.20.2
+  file: ^7.0.1
   path: ^1.8.0
   path_provider: ^2.0.5
-  gallery_saver: ^2.3.1
-  share_plus: ^6.3.2
-  youtube_player_flutter: ^8.0.0
-  url_launcher: ^6.0.12
+  saver_gallery: ^4.0.1
+  share_plus: ^10.1.4
+  # youtube_player_flutter: ^9.0.4
+  # youtube_player_iframe: ^5.1.5
+  url_launcher: ^6.3.1
   loading_animations: ^2.2.0
-  package_info_plus: ^3.1.0
-  lottie: ^2.3.2
+  package_info_plus: ^8.2.1
+  lottie: ^3.0.0
   platform: ^3.1.0
 
   # The following adds the Cupertino Icons font to your application.
