@@ -1,4 +1,4 @@
-package com.example.picture_of_the_day
+package com.barlow.picture_of_the_day
 
 import io.flutter.embedding.android.FlutterActivity
 
